Create new file with utils functions
